+++ src/main/java/uk/gov/digital/ho/hocs/info/api/TeamResource.java
@@ -2,11 +2,9 @@ package uk.gov.digital.ho.hocs.info.api;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import uk.gov.digital.ho.hocs.info.api.data.SimpleMapItem;
 import uk.gov.digital.ho.hocs.info.api.dto.*;
 import uk.gov.digital.ho.hocs.info.domain.model.Team;
 
-import java.util.List;
 import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
@@ -121,11 +119,6 @@ public class TeamResource {
     public ResponseEntity<TeamDto> getActiveTeams(@PathVariable UUID caseUUID, @PathVariable UUID topicUUID, @PathVariable String stageType) {
         Team team = teamService.getTeamByTopicAndStage(caseUUID, topicUUID, stageType);
         return ResponseEntity.ok(TeamDto.from(team));
-    }
-
-    @GetMapping(value = "/team/topic/stage/{stageType}")
-    public ResponseEntity<List<SimpleMapItem>> getTopicToTeamMappingByStageType(@PathVariable String stageType) {
-        return ResponseEntity.ok(teamService.getTopicToTeamMappingByStageType(stageType));
     }
 
     @GetMapping(value = "/teams/topic/{topicUUID}")
