+++ src/main/java/uk/gov/digital/ho/hocs/info/api/TopicTeamService.java
@@ -3,6 +3,7 @@ package uk.gov.digital.ho.hocs.info.api;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import uk.gov.digital.ho.hocs.info.api.data.SimpleMapItem;
 import uk.gov.digital.ho.hocs.info.api.dto.AddTeamToTopicDto;
 import uk.gov.digital.ho.hocs.info.client.audit.client.AuditClient;
 import uk.gov.digital.ho.hocs.info.domain.exception.ApplicationExceptions;
@@ -70,7 +71,7 @@ public class TopicTeamService {
 
         validateTeamTopicStageAndCase(teamUUID, topicUUID, stageType, caseType);
         TeamLink teamLink = Optional.ofNullable(teamLinkRepository.findByLinkValueAndLinkTypeAndCaseTypeAndStageType(topicUUID.toString(), "TOPIC", caseType, stageType))
-                 .orElse(new TeamLink(topicUUID.toString(), "TOPIC", teamUUID, caseType, stageType));
+                .orElse(new TeamLink(topicUUID.toString(), "TOPIC", teamUUID, caseType, stageType));
         teamLink.setResponsibleTeamUUID(teamUUID);
         teamLinkRepository.save(teamLink);
         log.info("Added team: {} to topic: {}", teamUUID, topicUUID);
@@ -78,18 +79,23 @@ public class TopicTeamService {
 
     }
 
-    private void validateTeamTopicStageAndCase(UUID teamUUID, UUID topicUUID, String stageType, String caseType){
+    public List<SimpleMapItem> getTopicToTeamMappingByStageType(String stageType) {
+        return teamRepository.findTopicToTeamMappingByStageType(stageType);
+    }
+
+    private void validateTeamTopicStageAndCase(UUID teamUUID, UUID topicUUID, String stageType, String caseType) {
         validateTopicUUID(topicUUID);
         validateStageType(stageType);
         validateCaseType(caseType);
         validateTeamUUID(teamUUID);
     }
 
-    private void validateTopicUUID(UUID topicUUID){
+    private void validateTopicUUID(UUID topicUUID) {
         Optional.ofNullable(topicService.getTopic(topicUUID))
                 .orElseThrow(() -> new ApplicationExceptions.TopicUpdateException
                         ("Unable to add team to topic, topic {} not found.", topicUUID));
     }
+
     private void validateStageType(String stageType) {
         Optional.ofNullable(stageTypeService.getStageType(stageType))
                 .orElseThrow(() -> new ApplicationExceptions.TopicUpdateException
@@ -102,11 +108,11 @@ public class TopicTeamService {
                         ("Unable to add team to topic, case type {} not found.", caseType));
     }
 
-    private void validateTeamUUID(UUID teamUUID){
+    private void validateTeamUUID(UUID teamUUID) {
         Team team = Optional.ofNullable(teamService.getTeam(teamUUID))
                 .orElseThrow(() -> new ApplicationExceptions.TopicUpdateException
                         ("Unable to add team to topic, topic already has team {} set for this stage and case type.", teamUUID));
-        if (!team.isActive()){
+        if (!team.isActive()) {
             throw new ApplicationExceptions.TopicUpdateException("Unable to add team to topic, team {} is inactive.", teamUUID);
         }
     }
