+++ src/test/java/uk/gov/digital/ho/hocs/info/api/TopicTeamResourceTest.java
@@ -7,10 +7,12 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import uk.gov.digital.ho.hocs.info.api.data.SimpleMapItem;
 import uk.gov.digital.ho.hocs.info.api.dto.*;
 import uk.gov.digital.ho.hocs.info.domain.model.TopicTeam;
 
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 import java.util.UUID;
 
@@ -49,8 +51,26 @@ public class TopicTeamResourceTest {
         UUID topicUUID = UUID.randomUUID();
 
         ResponseEntity response = topicTeamResource.addTeamToTopic(topicUUID, teamUUID, request);
-        verify(topicTeamService, times(1)).addTeamToTopic(topicUUID, teamUUID, request);
+        verify(topicTeamService).addTeamToTopic(topicUUID, teamUUID, request);
         verifyNoMoreInteractions(topicTeamService);
         assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
+    }
+
+    @Test
+    public void getTopicToTeamMappingByStageType() {
+        String testStageString = "STAGE_321";
+
+        List<SimpleMapItem> mockItems = List.of(mock(SimpleMapItem.class), mock(SimpleMapItem.class));
+        when(topicTeamService.getTopicToTeamMappingByStageType(testStageString)).thenReturn(mockItems);
+
+        ResponseEntity<List<SimpleMapItem>> response = topicTeamResource.getTopicToTeamMappingByStageType(testStageString);
+
+        assertThat(response).isNotNull();
+        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
+        assertThat(response.getBody()).isNotNull();
+        assertThat(response.getBody().size()).isEqualTo(2);
+
+        verify(topicTeamService).getTopicToTeamMappingByStageType(testStageString);
+        verifyNoMoreInteractions(topicTeamService);
     }
 }
