+++ src/test/java/uk/gov/digital/ho/hocs/info/api/TopicTeamServiceTest.java
@@ -5,6 +5,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
+import uk.gov.digital.ho.hocs.info.api.data.SimpleMapItem;
 import uk.gov.digital.ho.hocs.info.api.dto.AddTeamToTopicDto;
 import uk.gov.digital.ho.hocs.info.client.audit.client.AuditClient;
 import uk.gov.digital.ho.hocs.info.domain.exception.ApplicationExceptions;
@@ -13,7 +14,10 @@ import uk.gov.digital.ho.hocs.info.domain.repository.TeamLinkRepository;
 import uk.gov.digital.ho.hocs.info.domain.repository.TeamRepository;
 import uk.gov.digital.ho.hocs.info.domain.repository.TopicRepository;
 
-import java.util.*;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Set;
+import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
@@ -43,14 +47,14 @@ public class TopicTeamServiceTest {
 
     @Before
     public void setUp() {
-        this.topicTeamService = new TopicTeamService (teamLinkRepository,
-                     topicRepository,
-                     teamRepository,
-                     topicService,
-                     teamService,
-                     caseTypeService,
-                     stageTypeService,
-                     auditClient);
+        this.topicTeamService = new TopicTeamService(teamLinkRepository,
+                topicRepository,
+                teamRepository,
+                topicService,
+                teamService,
+                caseTypeService,
+                stageTypeService,
+                auditClient);
     }
 
     @Test
@@ -83,7 +87,7 @@ public class TopicTeamServiceTest {
 
     @Test
     public void shouldSuccessfullyAddTeamToTopic() {
-        AddTeamToTopicDto request = new AddTeamToTopicDto("MIN","DCU_MIN_MARKUP");
+        AddTeamToTopicDto request = new AddTeamToTopicDto("MIN", "DCU_MIN_MARKUP");
         UUID topicUUID = UUID.randomUUID();
         UUID teamUUID = UUID.randomUUID();
 
@@ -95,15 +99,15 @@ public class TopicTeamServiceTest {
         topicTeamService.addTeamToTopic(topicUUID, teamUUID, request);
 
 
-        verify(teamLinkRepository, times(1)).save(any());
-        verify(teamLinkRepository, times(1)).findByLinkValueAndLinkTypeAndCaseTypeAndStageType(any(), eq("TOPIC"), any(), any());
+        verify(teamLinkRepository).save(any());
+        verify(teamLinkRepository).findByLinkValueAndLinkTypeAndCaseTypeAndStageType(any(), eq("TOPIC"), any(), any());
         verifyNoMoreInteractions(teamLinkRepository);
     }
 
-    @Test (expected = ApplicationExceptions.TopicUpdateException.class)
+    @Test(expected = ApplicationExceptions.TopicUpdateException.class)
     public void shouldThrowExceptionWhenAddingNonExistentTeamToTopic() {
 
-        AddTeamToTopicDto request = new AddTeamToTopicDto("MIN","DCU_MIN_MARKUP");
+        AddTeamToTopicDto request = new AddTeamToTopicDto("MIN", "DCU_MIN_MARKUP");
         UUID topicUUID = UUID.randomUUID();
         UUID teamUUID = UUID.randomUUID();
         when(teamService.getTeam(teamUUID)).thenReturn(null);
@@ -117,7 +121,7 @@ public class TopicTeamServiceTest {
     @Test
     public void shouldNotAddTeamWhenAddingNonExistentTeamToTopic() {
 
-        AddTeamToTopicDto request = new AddTeamToTopicDto("MIN","DCU_MIN_MARKUP");
+        AddTeamToTopicDto request = new AddTeamToTopicDto("MIN", "DCU_MIN_MARKUP");
         UUID topicUUID = UUID.randomUUID();
         UUID teamUUID = UUID.randomUUID();
         when(teamService.getTeam(teamUUID)).thenReturn(null);
@@ -125,7 +129,8 @@ public class TopicTeamServiceTest {
         when(caseTypeService.getCaseType("MIN")).thenReturn(new CaseType());
         when(topicService.getTopic(topicUUID)).thenReturn(new Topic());
 
-        try { topicTeamService.addTeamToTopic(topicUUID, teamUUID, request);
+        try {
+            topicTeamService.addTeamToTopic(topicUUID, teamUUID, request);
         } catch (ApplicationExceptions.TopicUpdateException e) {
             // Do nothing.
         }
@@ -133,10 +138,10 @@ public class TopicTeamServiceTest {
         verifyZeroInteractions(teamLinkRepository);
     }
 
-    @Test (expected = ApplicationExceptions.TopicUpdateException.class)
+    @Test(expected = ApplicationExceptions.TopicUpdateException.class)
     public void shouldThrowExceptionWhenAddingInactiveTeamToTopic() {
 
-        AddTeamToTopicDto request = new AddTeamToTopicDto("MIN","DCU_MIN_MARKUP");
+        AddTeamToTopicDto request = new AddTeamToTopicDto("MIN", "DCU_MIN_MARKUP");
         UUID topicUUID = UUID.randomUUID();
         UUID teamUUID = UUID.randomUUID();
         when(teamService.getTeam(teamUUID)).thenReturn(new Team("name", false));
@@ -150,7 +155,7 @@ public class TopicTeamServiceTest {
     @Test
     public void shouldNotAddTeamWhenAddingInactiveTeamToTopic() {
 
-        AddTeamToTopicDto request = new AddTeamToTopicDto("MIN","DCU_MIN_MARKUP");
+        AddTeamToTopicDto request = new AddTeamToTopicDto("MIN", "DCU_MIN_MARKUP");
         UUID topicUUID = UUID.randomUUID();
         UUID teamUUID = UUID.randomUUID();
         when(teamService.getTeam(teamUUID)).thenReturn(new Team("name", false));
@@ -158,7 +163,8 @@ public class TopicTeamServiceTest {
         when(caseTypeService.getCaseType("MIN")).thenReturn(new CaseType());
         when(topicService.getTopic(topicUUID)).thenReturn(new Topic());
 
-        try { topicTeamService.addTeamToTopic(topicUUID, teamUUID, request);
+        try {
+            topicTeamService.addTeamToTopic(topicUUID, teamUUID, request);
         } catch (ApplicationExceptions.TopicUpdateException e) {
             // Do nothing.
         }
@@ -166,13 +172,13 @@ public class TopicTeamServiceTest {
         verifyZeroInteractions(teamLinkRepository);
     }
 
-    @Test (expected = ApplicationExceptions.TopicUpdateException.class)
+    @Test(expected = ApplicationExceptions.TopicUpdateException.class)
     public void shouldThrowExceptionWhenAddingTeamToNonexistentTopic() {
 
-        AddTeamToTopicDto request = new AddTeamToTopicDto("MIN","DCU_MIN_MARKUP");
+        AddTeamToTopicDto request = new AddTeamToTopicDto("MIN", "DCU_MIN_MARKUP");
         UUID topicUUID = UUID.randomUUID();
         UUID teamUUID = UUID.randomUUID();
-        
+
         topicTeamService.addTeamToTopic(topicUUID, teamUUID, request);
     }
 
@@ -192,10 +198,10 @@ public class TopicTeamServiceTest {
         verifyZeroInteractions(teamLinkRepository);
     }
 
-    @Test (expected = ApplicationExceptions.TopicUpdateException.class)
+    @Test(expected = ApplicationExceptions.TopicUpdateException.class)
     public void shouldThrowExceptionWhenAddingTeamToTopicAndCaseTypeIsInvalid() {
 
-        AddTeamToTopicDto request = new AddTeamToTopicDto("MIN","DCU_MIN_MARKUP");
+        AddTeamToTopicDto request = new AddTeamToTopicDto("MIN", "DCU_MIN_MARKUP");
         UUID topicUUID = UUID.randomUUID();
         UUID teamUUID = UUID.randomUUID();
 
@@ -208,12 +214,13 @@ public class TopicTeamServiceTest {
     @Test
     public void shouldNotAddTeamWhenAddingTeamToTopicAndCaseTypeIsInvalid() {
 
-        AddTeamToTopicDto request = new AddTeamToTopicDto("MIN","DCU_MIN_MARKUP");
+        AddTeamToTopicDto request = new AddTeamToTopicDto("MIN", "DCU_MIN_MARKUP");
         UUID topicUUID = UUID.randomUUID();
         UUID teamUUID = UUID.randomUUID();
         when(stageTypeService.getStageType("DCU_MIN_MARKUP")).thenReturn(new StageTypeEntity());
         when(topicService.getTopic(topicUUID)).thenReturn(new Topic());
-        try { topicTeamService.addTeamToTopic(topicUUID, teamUUID, request);
+        try {
+            topicTeamService.addTeamToTopic(topicUUID, teamUUID, request);
         } catch (ApplicationExceptions.TopicUpdateException e) {
             // Do nothing.
         }
@@ -221,10 +228,10 @@ public class TopicTeamServiceTest {
         verifyZeroInteractions(teamLinkRepository);
     }
 
-    @Test (expected = ApplicationExceptions.TopicUpdateException.class)
+    @Test(expected = ApplicationExceptions.TopicUpdateException.class)
     public void shouldThrowExceptionWhenAddingTeamToTopicAndStageTypeIsInvalid() {
 
-        AddTeamToTopicDto request = new AddTeamToTopicDto("MIN","DCU_MIN_MARKUP");
+        AddTeamToTopicDto request = new AddTeamToTopicDto("MIN", "DCU_MIN_MARKUP");
         UUID topicUUID = UUID.randomUUID();
         UUID teamUUID = UUID.randomUUID();
         when(topicService.getTopic(topicUUID)).thenReturn(new Topic());
@@ -235,11 +242,12 @@ public class TopicTeamServiceTest {
     @Test
     public void shouldNotAddTeamWhenAddingTeamToTopicAndStageTypeIsInvalid() {
 
-        AddTeamToTopicDto request = new AddTeamToTopicDto("MIN","DCU_MIN_MARKUP");
+        AddTeamToTopicDto request = new AddTeamToTopicDto("MIN", "DCU_MIN_MARKUP");
         UUID topicUUID = UUID.randomUUID();
         UUID teamUUID = UUID.randomUUID();
         when(topicService.getTopic(topicUUID)).thenReturn(new Topic());
-        try { topicTeamService.addTeamToTopic(topicUUID, teamUUID, request);
+        try {
+            topicTeamService.addTeamToTopic(topicUUID, teamUUID, request);
         } catch (ApplicationExceptions.TopicUpdateException e) {
             // Do nothing.
         }
@@ -250,7 +258,7 @@ public class TopicTeamServiceTest {
     @Test
     public void shouldUpdateTeamWhenAddingTeamToTopicAndTopicAlreadyHasATeam() {
 
-        AddTeamToTopicDto request = new AddTeamToTopicDto("MIN","DCU_MIN_MARKUP");
+        AddTeamToTopicDto request = new AddTeamToTopicDto("MIN", "DCU_MIN_MARKUP");
         UUID topicUUID = UUID.randomUUID();
         UUID teamUUID = UUID.randomUUID();
 
@@ -263,16 +271,16 @@ public class TopicTeamServiceTest {
 
         topicTeamService.addTeamToTopic(topicUUID, teamUUID, request);
 
-        verify(teamLinkRepository, times(1)).findByLinkValueAndLinkTypeAndCaseTypeAndStageType(any(), eq("TOPIC"), any(), any());
-        verify(teamLinkRepository, times(1)).save(any());
+        verify(teamLinkRepository).findByLinkValueAndLinkTypeAndCaseTypeAndStageType(any(), eq("TOPIC"), any(), any());
+        verify(teamLinkRepository).save(any());
         verifyNoMoreInteractions(teamLinkRepository);
     }
 
 
     @Test
-    public void shouldAuditAddTeamToTopic(){
+    public void shouldAuditAddTeamToTopic() {
 
-        AddTeamToTopicDto request = new AddTeamToTopicDto("MIN","DCU_MIN_MARKUP");
+        AddTeamToTopicDto request = new AddTeamToTopicDto("MIN", "DCU_MIN_MARKUP");
         UUID topicUUID = UUID.randomUUID();
         UUID teamUUID = UUID.randomUUID();
 
@@ -283,7 +291,23 @@ public class TopicTeamServiceTest {
 
         topicTeamService.addTeamToTopic(topicUUID, teamUUID, request);
 
-        verify(auditClient, times(1)).addTeamToTopicAudit(any());
+        verify(auditClient).addTeamToTopicAudit(any());
         verifyNoMoreInteractions(auditClient);
+    }
+
+    @Test
+    public void getTopicToTeamMappingByStageType() {
+        String testStageString = "STAGE_321";
+
+        List<SimpleMapItem> mockItems = List.of(mock(SimpleMapItem.class), mock(SimpleMapItem.class));
+        when(teamRepository.findTopicToTeamMappingByStageType(testStageString)).thenReturn(mockItems);
+
+        List<SimpleMapItem> results = topicTeamService.getTopicToTeamMappingByStageType(testStageString);
+
+        assertThat(results).isNotNull();
+        assertThat(results.size()).isEqualTo(2);
+
+        verify(teamRepository).findTopicToTeamMappingByStageType(testStageString);
+        verifyNoMoreInteractions(teamRepository);
     }
 }
